LPS-75792 Move Statistics and SummaryStatistics into internal package
