+++ modules/apps/foundation/portal/portal-monitoring/src/main/java/com/liferay/portal/monitoring/internal/statistics/BaseStatistics.java
@@ -14,8 +14,6 @@
 
 package com.liferay.portal.monitoring.internal.statistics;
 
-import com.liferay.portal.monitoring.statistics.Statistics;
-
 /**
  * @author Rajesh Thiagarajan
  * @author Brian Wing Shun Chan
